requirements.txt
pdfplumber==0.11.0
camelot-py==0.11.0
pandas==2.2.2
openpyxl==3.1.2
ghostscript
